+++ buildout/src/Euphorie/euphorie/deployment/upgrade/v2.py
@@ -1,10 +1,10 @@
 import logging
 from Acquisition import aq_parent
-from Products.CMFCore.utils import getToolByName
 
 log = logging.getLogger(__name__)
 
 def resetSurveyWorkflow(context):
+    from Products.CMFCore.utils import getToolByName
     siteroot=aq_parent(context)
     wt=getToolByName(siteroot, 'portal_workflow')
     workflow=wt.survey
@@ -18,4 +18,13 @@ def resetSurveyWorkflow(context):
     count=wt._recursiveUpdateRoleMappings(siteroot, {"survey": workflow})
     log.info("Updated permissions for %d objects", count)
 
+
+def resetPublishPermission(context):
+    from AccessControl.Permission import Permission
+
+    siteroot=aq_parent(context)
+    permission=Permission("Euphorie: Publish a Survey", (), siteroot)
+    if "CountryManager" not in permission.getRoles(default=[]):
+        permission.setRole("CountryManager", True)
+        log.info("Adding publish permission for country managers")